// src/models/AirtableDtos/PropertyDto.ts
export class PropertyDto {
    id: number = 0;
    name: string = "";
    externalName: string = "";
    internalCode: string = "";
    active: number = 0;
    checkIn: string = "";
    checkOut: string = "";
    latitude: number = 0;
    longitude: number = 0;
    street1: string = "";
    city: string = "";
    alphaCode: string = "";
    postalCode: string = "";
    thumbnail_url: string = "";
}